[nfc] Rename binary_sensor source files
