+++ esphome/components/nfc/binary_sensor/nfc_binary_sensor.h
